make some updates

+++ src/common/multer/deleteFile.ts
@@ -1,5 +1,7 @@
+import { existsSync } from 'fs';
 import { unlink } from 'fs/promises'
 
 export const deleteFile = async (fullPath: string) => {
-    await unlink(fullPath)
-}
+    if (!existsSync(fullPath)) return;
+    await unlink(fullPath);
+  };
